Minor refactor of lightprobe related code
